api controller/index updated
